+++ kondo-lib/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "kondo-lib"
+version = "0.1.0"
+authors = ["Trent Billington <trent.h.billington@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+walkdir = "2"
